Use legacy hubness reduction implementation for tests only
