+++ skhubness/reduction/tests/test_dis_sim.py
@@ -5,8 +5,8 @@ from sklearn.datasets import make_classification
 from sklearn.utils.testing import assert_array_almost_equal
 from sklearn.utils.testing import assert_array_equal
 from sklearn.utils.testing import assert_raises
-from skhubness.reduction import DisSimLocal
 from skhubness.neighbors import NearestNeighbors, KNeighborsClassifier
+from skhubness.reduction.tests.reference_algorithms import DisSimLocal
 
 
 def test_squared():
